feat: 페르소나 이미지 어댑터 수정 - character 리스트를 어댑터 내부에 선언하여 중복된 코드 삭제
