+++ app/src/main/java/umc/cozymate/ui/my_page/my_info/UpdateCharacterFragment.kt
@@ -64,25 +64,7 @@ class UpdateCharacterFragment: Fragment(), CharacterItemClickListener {
     }
 
     private fun initCharacterList() {
-        val characters = listOf(
-            CharacterItem(R.drawable.character_id_1),
-            CharacterItem(R.drawable.character_id_2),
-            CharacterItem(R.drawable.character_id_3),
-            CharacterItem(R.drawable.character_id_5),
-            CharacterItem(R.drawable.character_id_6),
-            CharacterItem(R.drawable.character_id_4),
-            CharacterItem(R.drawable.character_id_15),
-            CharacterItem(R.drawable.character_id_14),
-            CharacterItem(R.drawable.character_id_8),
-            CharacterItem(R.drawable.character_id_7),
-            CharacterItem(R.drawable.character_id_11),
-            CharacterItem(R.drawable.character_id_12),
-            CharacterItem(R.drawable.character_id_10),
-            CharacterItem(R.drawable.character_id_13),
-            CharacterItem(R.drawable.character_id_9),
-            CharacterItem(R.drawable.character_id_16),
-        )
-        val adapter = CharactersAdapter(characters, this)
+        val adapter = CharactersAdapter(this)
         binding.rvList.adapter = adapter
         binding.rvList.run {
             layoutManager = GridLayoutManager(context, 4)
