+++ app/src/main/java/umc/cozymate/ui/onboarding/adapter/CharactersAdapter.kt
@@ -3,14 +3,30 @@ package umc.cozymate.ui.onboarding.adapter
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import umc.cozymate.R
 import umc.cozymate.databinding.ItemCharacterBinding
 
 class CharactersAdapter(
-    private var characters: List<CharacterItem>,
     private var onItemClickListener: CharacterItemClickListener
-) :
-    RecyclerView.Adapter<CharacterViewHolder>() {
-
+) : RecyclerView.Adapter<CharacterViewHolder>() {
+    val characters = listOf(
+        CharacterItem(R.drawable.character_id_1),
+        CharacterItem(R.drawable.character_id_2),
+        CharacterItem(R.drawable.character_id_3),
+        CharacterItem(R.drawable.character_id_5),
+        CharacterItem(R.drawable.character_id_6),
+        CharacterItem(R.drawable.character_id_4),
+        CharacterItem(R.drawable.character_id_15),
+        CharacterItem(R.drawable.character_id_14),
+        CharacterItem(R.drawable.character_id_8),
+        CharacterItem(R.drawable.character_id_7),
+        CharacterItem(R.drawable.character_id_11),
+        CharacterItem(R.drawable.character_id_12),
+        CharacterItem(R.drawable.character_id_10),
+        CharacterItem(R.drawable.character_id_13),
+        CharacterItem(R.drawable.character_id_9),
+        CharacterItem(R.drawable.character_id_16),
+    )
     private var selectedPosition: Int = -1
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CharacterViewHolder {
@@ -33,10 +49,5 @@ class CharactersAdapter(
             }
             onItemClickListener.onItemClick(characters[position], selectedPosition)
         }
-    }
-
-    fun setData(data: List<CharacterItem>) {
-        characters = data
-        notifyDataSetChanged()
     }
 }
